Limit needless recalculation of symbolic transformations.
Use a hashtable to store already processed values in
various places, and thus avoid needlessy rescanning them.
The rescans were getting rather slow even with a single
iref in the source code.

Note that the iref macro may be called multiple times
if a wrapping macro call clones the code.