Generalize index loop generation code.
Introduce an additional form 'loop-range' that represents
an index iteration in a general form, pull wrapping code
out of loop-indexes into a helper function, and fix
replace-unquoted to handle let and let* properly.

Also use a special variable to pass the layer tag to
index-iterexpr. It probably needs some refactoring.