Implement type inference and annotation insertion.
In order to force ECL to generate native floating-point
code instead of manipulating boxed objects, it is necessary
to fill the expressions with type annotations. This
requires performing simple type inference (assuming
that there are only float, integer and boolean types),
and postprocessing the expression.

I also add a new 'calc' macro that is like letv, but
with annotations.