Optimize expressions before compilation.
Intel Core 2 CPU can perform one multiplication and
addition every clock tick, provided that they are
independent. The dependent latency is 5 clocks for
multiplication, and 3 clocks for addition; division
is ten times slower.

This commit adds some code that restructures the
expression tree to make it more balanced, and minimize
the number of divisions and pointless negations.