Avoid deadlocks in the thread manager.
If the main thread posts a new task and signals the
condition before the worker thread waits on it, it
would miss the signal. Thus, it should not wait if
the task is set.

On the other hand, it is necessary to prevent it from
picking up the same task twice, thus a new condition.