Ensure that assignments within a (compute...) are independent.
Force the code motion system to factor away read accesses to
the same addresses that are written to in the statement. Thus
it is guaranteed that assignments cannot affect each other
no matter the order. Note that code motion already caused
this effect in some cases, so this only makes it consistent.