Implement splicing of the innermost directed loop.
Extract as much code as possible from the directed
part of the loop. Includes detection of carried
dependencies, which may mark each node as depending
on a carried value (high), or affecting such value
(low). All nodes with both marks must be in the
directed part, while others may be moved out.