Use a structure for the part of the range that must be shared.
The refactoring code destructively modifies some of the index
range parameters, so they must remain shared throughout the
expression restructuring passes. When the ranging spec is a
list, ensuring this is difficult.

This changes the mutable part of the range specification into
a structure, providing macros to easily create and match such
nodes. Code throughout the library is switched to the new format.