Implement splitting of flat trees by common subsets.
Add code to determine that e.g. (+ a b c d) and
(+ a c d) have a common subpart, and split it off
so that it can be further exploited by the common
subexpression elimination pass.

Works by doing a O(N^2) intersection of all similar
terms, and then extracting subsets in a certain order
of precedence.